Fix Function Calling Reference Issues
I don't really like the way I made this so I will need to revisit it.

Essentially, the function body needs to be referenced when trying to execute it (because the environment owns the variable); this forces me to use references everywhere block expressions are used, and thus everywhere expressions & statements are used.

However, this leads to a dilemma where the evaluator encounters the FunctionLiteral and attempts to create a Function object with it but it needs an Rc reference to the BlockExpression, which requires ownership. Yet the literal is passed by reference so it's not possible to get ownership.

The solution to this would be either to make the FunctionLiteral use Rc<BlockExpression> in the first place (what I ended up doing), or to use &references instead of Rc, or to clone the BlockExpression when creating the Function object.

I didn't like the idea of cloning or of overhauling the entire code, so I just made the AST create an Rc in the first place. However, this feels messy and breaks down the separation of concerns between the parser and evaluator, which will need to be revisited.